Moving more files around... now compiling a common-core library and added STM32G491 and STM32F303 common-module code directories.
