+++ stm32-modules/common/STM32F303/gdbinit.template
