+++ stm32-modules/common/STM32G491/gdbinit.template
