+++ stm32-modules/common/STM32G491/startup_stm32g491vetx.s
