+++ stm32-modules/common/STM32G491/stm32g4discovery.cfg
