+++ stm32-modules/common/simulator/cli_parser.cpp
