+++ stm32-modules/common/simulator/stdin_sim_driver.cpp
