+++ stm32-modules/common/src/CMakeLists.txt
@@ -6,35 +6,26 @@
 find_package(Python)
 find_package(Git QUIET)
 
-# This is like version for project but works when there aren't any tags yet
-execute_process(COMMAND ${GIT_EXECUTABLE} describe --tags --match "${TARGET_MODULE_NAME}*"
-  WORKING_DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}"
-  OUTPUT_VARIABLE TAGNAME
-  OUTPUT_STRIP_TRAILING_WHITESPACE)
-if (NOT TAGNAME)
-  set(TAGNAME "@(dev)-")
-endif()
-
-string(REGEX MATCH "@.*$" PREPENDED_VERSION ${TAGNAME})
-string(SUBSTRING "${PREPENDED_VERSION}" 1 -1 VERSION)
-if (NOT VERSION)
-  set(VERSION "(dev)")
-endif()
-execute_process(COMMAND ${GIT_EXECUTABLE} rev-parse --short HEAD
-  WORKING_DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}"
-  OUTPUT_VARIABLE HASH
-  OUTPUT_STRIP_TRAILING_WHITESPACE)
-
-set(${TARGET_MODULE_NAME}_VERSION "${VERSION}-${HASH}")
-
-configure_file(./version.cpp.in ./version.cpp)
+add_custom_command(
+  COMMAND Python::Interpreter ${CMAKE_CURRENT_SOURCE_DIR}/generate_thermistor_table.py
+    ${CMAKE_CURRENT_SOURCE_DIR}/ntcg104ed104dtdsx.csv
+    ${CMAKE_CURRENT_BINARY_DIR}/thermistor_lookups.hpp
+    ${CMAKE_CURRENT_BINARY_DIR}/thermistor_lookups.cpp
+  DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/generate_thermistor_table.py
+          ${CMAKE_CURRENT_SOURCE_DIR}/ntcg104ed104dtdsx.csv
+  OUTPUT ${CMAKE_CURRENT_BINARY_DIR}/thermistor_lookups.hpp
+         ${CMAKE_CURRENT_BINARY_DIR}/thermistor_lookups.cpp
+  WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}
+  )
 
 # Configure lintable/nonlintable sources here
 set(CORE_LINTABLE_SOURCES
-  ${CMAKE_CURRENT_SOURCE_DIR}/errors.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/pid.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/thermistor_conversion.cpp
   )
-set(CORE_NONLINTABLE_SOURCES
-  ${CMAKE_CURRENT_BINARY_DIR}/version.cpp)
+set(CORE_NONLINTABLE_SOURCES 
+  ${CMAKE_CURRENT_BINARY_DIR}/thermistor_lookups.cpp
+  )
 
 add_library(${TARGET_MODULE_NAME}-core STATIC
   ${CORE_LINTABLE_SOURCES}
