+++ stm32-modules/common/src/generate_thermistor_table.py
