+++ stm32-modules/heater-shaker/src/CMakeLists.txt
@@ -28,26 +28,11 @@ set(heater-shaker_VERSION "${VERSION}-${HASH}")
 
 configure_file(./version.cpp.in ./version.cpp)
 
-add_custom_command(
-  COMMAND Python::Interpreter ${CMAKE_CURRENT_SOURCE_DIR}/generate_thermistor_table.py
-    ${CMAKE_CURRENT_SOURCE_DIR}/ntcg104ed104dtdsx.csv
-    ${CMAKE_CURRENT_BINARY_DIR}/thermistor_lookups.hpp
-    ${CMAKE_CURRENT_BINARY_DIR}/thermistor_lookups.cpp
-  DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/generate_thermistor_table.py
-          ${CMAKE_CURRENT_SOURCE_DIR}/ntcg104ed104dtdsx.csv
-  OUTPUT ${CMAKE_CURRENT_BINARY_DIR}/thermistor_lookups.hpp
-         ${CMAKE_CURRENT_BINARY_DIR}/thermistor_lookups.cpp
-  WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}
-  )
-
 set(CORE_LINTABLE_SOURCES
-  ${CMAKE_CURRENT_SOURCE_DIR}/pid.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/errors.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/thermistor_conversion.cpp
   )
 
 set(CORE_NONLINTABLE_SOURCES
-  ${CMAKE_CURRENT_BINARY_DIR}/thermistor_lookups.cpp
   ${CMAKE_CURRENT_BINARY_DIR}/version.cpp)
 
 add_library(heater-shaker-core STATIC
@@ -62,7 +47,7 @@ set_target_properties(heater-shaker-core
 target_include_directories(heater-shaker-core
   PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/../../include/heater-shaker
   PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/../../include/common
-  PRIVATE ${CMAKE_CURRENT_BINARY_DIR}  # need to add this to get the generated conversion header
+  PRIVATE ${CMAKE_CURRENT_BINARY_DIR}../../common/src  # need to add this to get the generated conversion header
   )
 
 target_compile_options(heater-shaker-core
