--- stm32-modules/heater-shaker/src/ntcg104ed104dtdsx.csv
@@ -1,172 +1,0 @@
-P/N,NTCG104ED104DTDSX,,,,,
-R25,100,K +/-,1,%,,
-B25/50,4250,K  +/- ,1,%,,
- ,,,,,,
-Temp,Min,Nom,Max,B25/T, -dT,dT
-(degC),(k ohm),(k ohm),(k ohm),(K),(degC),(degC)
--40,4151,4251,4353,4010,-0.3,0.3
--39,3871,3962,4056,4014,-0.3,0.3
--38,3611,3695,3781,4017,-0.3,0.3
--37,3370,3447,3526,4020,-0.3,0.3
--36,3147,3218,3290,4024,-0.3,0.3
--35,2939,3005,3072,4027,-0.3,0.3
--34,2747,2807,2869,4030,-0.3,0.3
--33,2569,2624,2681,4034,-0.3,0.3
--32,2403,2454,2506,4037,-0.3,0.3
--31,2249,2296,2344,4040,-0.3,0.3
--30,2106,2149,2193,4043,-0.3,0.3
--29,1972,2012,2053,4047,-0.3,0.3
--28,1848,1885,1923,4050,-0.3,0.3
--27,1733,1767,1801,4053,-0.3,0.3
--26,1625,1656,1688,4056,-0.3,0.3
--25,1525,1554,1583,4059,-0.3,0.3
--24,1431,1458,1485,4062,-0.3,0.3
--23,1344,1369,1394,4066,-0.3,0.3
--22,1263,1286,1309,4069,-0.3,0.3
--21,1187,1208,1229,4072,-0.3,0.3
--20,1116,1135,1155,4075,-0.3,0.3
--19,1050,1068,1086,4078,-0.3,0.3
--18,987.7,1004,1021,4081,-0.3,0.3
--17,929.7,945,960.4,4084,-0.3,0.3
--16,875.5,889.6,903.9,4087,-0.3,0.3
--15,824.8,837.8,851,4090,-0.3,0.3
--14,777.3,789.3,801.5,4093,-0.3,0.3
--13,732.8,743.9,755.1,4096,-0.2,0.2
--12,691.1,701.3,711.7,4099,-0.2,0.2
--11,652,661.5,671.1,4102,-0.2,0.2
--10,615.3,624.1,633,4105,-0.2,0.2
--9,580.9,589,597.3,4108,-0.2,0.2
--8,548.6,556.2,563.7,4111,-0.2,0.2
--7,518.3,525.3,532.3,4113,-0.2,0.2
--6,489.9,496.3,502.8,4116,-0.2,0.2
--5,463.2,469.1,475.1,4119,-0.2,0.2
--4,438,443.5,449.1,4122,-0.2,0.2
--3,414.4,419.5,424.6,4125,-0.2,0.2
--2,392.2,396.9,401.6,4127,-0.2,0.2
--1,371.3,375.6,380,4130,-0.2,0.2
-0,351.6,355.6,359.7,4133,-0.2,0.2
-1,333.1,336.8,340.6,4136,-0.2,0.2
-2,315.7,319.1,322.5,4138,-0.2,0.2
-3,299.2,302.4,305.6,4141,-0.2,0.2
-4,283.7,286.7,289.6,4144,-0.2,0.2
-5,269.1,271.8,274.6,4147,-0.2,0.2
-6,255.3,257.8,260.4,4149,-0.2,0.2
-7,242.4,244.7,247,4152,-0.2,0.2
-8,230.1,232.2,234.4,4154,-0.2,0.2
-9,218.5,220.5,222.5,4157,-0.2,0.2
-10,207.6,209.4,211.2,4160,-0.2,0.2
-11,197.3,198.9,200.6,4162,-0.2,0.2
-12,187.5,189,190.6,4165,-0.2,0.2
-13,178.3,179.7,181.1,4167,-0.2,0.2
-14,169.6,170.9,172.2,4170,-0.2,0.2
-15,161.3,162.5,163.7,4172,-0.1,0.1
-16,153.5,154.6,155.7,4175,-0.1,0.1
-17,146.1,147.2,148.2,4177,-0.1,0.1
-18,139.1,140.1,141,4180,-0.1,0.1
-19,132.5,133.4,134.2,4182,-0.1,0.1
-20,126.3,127,127.8,4185,-0.1,0.1
-21,120.3,121,121.8,4187,-0.1,0.1
-22,114.7,115.4,116,4190,-0.1,0.1
-23,109.4,110,110.6,4192,-0.1,0.1
-24,104.3,104.8,105.4,4194,-0.1,0.1
-25,99.5,100,100.5,4197,-0.1,0.1
-26,94.9,95.4,95.9,4199,-0.1,0.1
-27,90.54,91.04,91.54,4201,-0.1,0.1
-28,86.4,86.9,87.39,4204,-0.1,0.1
-29,82.47,82.97,83.46,4206,-0.1,0.1
-30,78.74,79.23,79.72,4208,-0.1,0.1
-31,75.2,75.69,76.17,4210,-0.1,0.1
-32,71.84,72.32,72.8,4213,-0.1,0.1
-33,68.64,69.12,69.59,4215,-0.2,0.2
-34,65.6,66.07,66.54,4217,-0.2,0.2
-35,62.72,63.18,63.64,4219,-0.2,0.2
-36,59.97,60.42,60.88,4221,-0.2,0.2
-37,57.36,57.81,58.25,4223,-0.2,0.2
-38,54.87,55.31,55.76,4226,-0.2,0.2
-39,52.51,52.94,53.38,4228,-0.2,0.2
-40,50.26,50.68,51.11,4230,-0.2,0.2
-41,48.12,48.53,48.95,4232,-0.2,0.2
-42,46.08,46.49,46.9,4234,-0.2,0.2
-43,44.13,44.53,44.94,4236,-0.2,0.2
-44,42.28,42.67,43.07,4238,-0.2,0.2
-45,40.52,40.9,41.29,4240,-0.2,0.2
-46,38.83,39.21,39.59,4242,-0.2,0.2
-47,37.23,37.6,37.97,4244,-0.2,0.2
-48,35.7,36.06,36.43,4246,-0.2,0.2
-49,34.24,34.6,34.95,4248,-0.3,0.3
-50,32.85,33.19,33.55,4250,-0.3,0.3
-51,31.52,31.86,32.2,4252,-0.3,0.3
-52,30.25,30.58,30.92,4254,-0.3,0.3
-53,29.04,29.36,29.69,4256,-0.3,0.3
-54,27.88,28.2,28.52,4258,-0.3,0.3
-55,26.78,27.09,27.4,4259,-0.3,0.3
-56,25.72,26.03,26.33,4261,-0.3,0.3
-57,24.71,25.01,25.31,4263,-0.3,0.3
-58,23.75,24.04,24.33,4265,-0.3,0.3
-59,22.83,23.11,23.4,4267,-0.3,0.3
-60,21.95,22.22,22.5,4269,-0.3,0.3
-61,21.1,21.37,21.65,4270,-0.3,0.3
-62,20.3,20.56,20.83,4272,-0.3,0.3
-63,19.52,19.78,20.04,4274,-0.3,0.3
-64,18.78,19.04,19.29,4276,-0.4,0.4
-65,18.08,18.32,18.57,4277,-0.4,0.4
-66,17.4,17.64,17.88,4279,-0.4,0.4
-67,16.75,16.99,17.22,4281,-0.4,0.4
-68,16.13,16.36,16.59,4282,-0.4,0.4
-69,15.54,15.76,15.99,4284,-0.4,0.4
-70,14.97,15.18,15.4,4286,-0.4,0.4
-71,14.42,14.63,14.85,4287,-0.4,0.4
-72,13.9,14.1,14.31,4289,-0.4,0.4
-73,13.39,13.6,13.8,4290,-0.4,0.4
-74,12.91,13.11,13.31,4292,-0.4,0.4
-75,12.45,12.64,12.84,4293,-0.4,0.4
-76,12.01,12.19,12.39,4295,-0.4,0.4
-77,11.58,11.76,11.95,4296,-0.4,0.4
-78,11.17,11.35,11.53,4298,-0.5,0.5
-79,10.78,10.96,11.13,4299,-0.5,0.5
-80,10.41,10.58,10.75,4301,-0.5,0.5
-81,10.04,10.21,10.38,4302,-0.5,0.5
-82,9.697,9.859,10.02,4304,-0.5,0.5
-83,9.364,9.522,9.683,4305,-0.5,0.5
-84,9.044,9.198,9.355,4307,-0.5,0.5
-85,8.736,8.887,9.04,4308,-0.5,0.5
-86,8.44,8.587,8.737,4309,-0.5,0.5
-87,8.156,8.299,8.445,4311,-0.5,0.5
-88,7.882,8.022,8.165,4312,-0.5,0.5
-89,7.619,7.756,7.895,4313,-0.5,0.5
-90,7.367,7.5,7.636,4315,-0.5,0.5
-91,7.123,7.254,7.386,4316,-0.6,0.6
-92,6.889,7.016,7.146,4317,-0.6,0.6
-93,6.664,6.788,6.914,4319,-0.6,0.6
-94,6.447,6.568,6.692,4320,-0.6,0.6
-95,6.238,6.357,6.477,4321,-0.6,0.6
-96,6.038,6.153,6.271,4322,-0.6,0.6
-97,5.844,5.957,6.072,4323,-0.6,0.6
-98,5.658,5.768,5.88,4325,-0.6,0.6
-99,5.478,5.586,5.695,4326,-0.6,0.6
-100,5.305,5.41,5.517,4327,-0.6,0.6
-101,5.138,5.241,5.345,4328,-0.6,0.6
-102,4.978,5.078,5.18,4329,-0.6,0.6
-103,4.823,4.921,5.02,4330,-0.7,0.7
-104,4.674,4.769,4.866,4331,-0.7,0.7
-105,4.53,4.623,4.718,4332,-0.7,0.7
-106,4.391,4.482,4.575,4334,-0.7,0.7
-107,4.257,4.346,4.437,4335,-0.7,0.7
-108,4.128,4.215,4.303,4336,-0.7,0.7
-109,4.003,4.088,4.175,4337,-0.7,0.7
-110,3.883,3.966,4.05,4338,-0.7,0.7
-111,3.767,3.848,3.931,4339,-0.7,0.7
-112,3.655,3.734,3.815,4340,-0.7,0.7
-113,3.546,3.624,3.703,4340,-0.7,0.7
-114,3.442,3.518,3.595,4341,-0.7,0.8
-115,3.341,3.415,3.491,4342,-0.8,0.8
-116,3.244,3.316,3.39,4343,-0.8,0.8
-117,3.149,3.22,3.292,4344,-0.8,0.8
-118,3.059,3.128,3.198,4345,-0.8,0.8
-119,2.971,3.038,3.107,4346,-0.8,0.8
-120,2.886,2.952,3.019,4347,-0.8,0.8
-121,2.804,2.868,2.934,4347,-0.8,0.8
-122,2.724,2.787,2.852,4348,-0.8,0.8
-123,2.647,2.709,2.772,4349,-0.8,0.8
-124,2.573,2.634,2.695,4350,-0.8,0.8
-125,2.501,2.561,2.621,4351,-0.8,0.9
