+++ stm32-modules/include/common/core/thermistor_conversion.hpp
