+++ stm32-modules/include/heater-shaker/heater-shaker/heater_task.hpp
@@ -9,11 +9,11 @@
 #include <variant>
 
 #include "core/pid.hpp"
+#include "core/thermistor_conversion.hpp"
 #include "hal/message_queue.hpp"
 #include "heater-shaker/errors.hpp"
 #include "heater-shaker/messages.hpp"
 #include "heater-shaker/tasks.hpp"
-#include "heater-shaker/thermistor_conversion.hpp"
 
 /* Need a forward declaration for this because of recursive includes */
 namespace tasks {
